backEnd/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.4.1=pypi_0
backcall=0.2.0=pypi_0
blas=2.120=mkl
blas-devel=3.9.0=20_win64_mkl
blinker=1.7.0=pypi_0
brotli-python=1.1.0=py38hd3f51b4_1
ca-certificates=2023.11.17=h56e8100_0
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.0=pypi_0
contourpy=1.1.1=pypi_0
cuda-cccl=12.3.101=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-opencl-dev=12.3.101=0
cuda-profiler-api=12.3.101=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
executing=2.0.1=pypi_0
filelock=3.13.1=pyhd8ed1ab_0
flask=3.0.0=pypi_0
flask-cors=4.0.0=pypi_0
fonttools=4.46.0=pypi_0
freetype=2.12.1=hdaf720e_2
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.0=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.2.0=h57928b3_50497
ipykernel=6.27.1=pypi_0
ipython=8.12.3=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
kiwisolver=1.4.5=pypi_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=20_win64_mkl
libcblas=3.9.0=20_win64_mkl
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.4.101=0
libcurand-dev=10.3.4.101=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.19=hcfcfb64_0
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=hcfcfb64_2
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=20_win64_mkl
liblapacke=3.9.0=20_win64_mkl
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h19919ed_0
libsqlite=3.44.2=hcfcfb64_0
libtiff=4.6.0=h6e2ebb7_2
libuv=1.44.2=hcfcfb64_1
libwebp-base=1.3.2=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.11.6=hc3477c8_0
libzlib=1.2.13=hcfcfb64_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.3=py38h91455d4_1
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.2.0=h6a75c08_50497
mkl-devel=2023.2.0=h57928b3_50497
mkl-include=2023.2.0=h6a75c08_50497
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
nest-asyncio=1.5.8=pypi_0
networkx=3.1=pyhd8ed1ab_0
numpy=1.24.4=py38h1d91fd2_0
openjpeg=2.5.0=h3d672ee_3
openssl=1.1.1w=hcfcfb64_0
packaging=23.2=pypi_0
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.1.0=py38hc375fad_0
pip=23.3.2=pyhd8ed1ab_0
platformdirs=4.1.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
psutil=5.9.7=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pure-eval=0.2.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pyh0701188_6
python=3.8.0=hc9e8b01_5
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=2.1.2=py3.8_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pywin32=306=pypi_0
pyyaml=6.0.1=py38h91455d4_1
pyzmq=25.1.2=pypi_0
requests=2.31.0=pyhd8ed1ab_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
sqlite=3.44.2=hcfcfb64_0
stack-data=0.6.3=pypi_0
sympy=1.12=pyh04b8f61_3
tbb=2021.11.0=h91493d7_0
tk=8.6.13=h5226925_1
torchaudio=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4=pypi_0
traitlets=5.14.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.1.0=pyhd8ed1ab_0
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wcwidth=0.2.12=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zipp=3.17.0=pypi_0
zstd=1.5.5=h12be248_0
